+++ wp-content/themes/vf-wp-groups/functions/groups-customize.php
@@ -1,0 +1,17 @@
+<?php
+
+if( ! defined( 'ABSPATH' ) ) exit;
+
+if ( ! class_exists('VF_Groups_Customize') ) :
+
+class VF_Groups_Customize {
+
+  public function __construct() {
+    // Nothing...
+  }
+
+} // VF_Groups_Customize
+
+endif;
+
+?>
