+++ wp-content/themes/vf-wp-groups/functions/groups-theme.php
@@ -1,0 +1,250 @@
+<?php
+
+if( ! defined( 'ABSPATH' ) ) exit;
+
+if ( ! class_exists('VF_Groups_Theme') ) :
+
+require_once('groups-customize.php');
+
+class VF_Groups_Theme {
+
+  private $theme_customize;
+
+  public function __construct() {
+    // Initialize sub-class instances
+    $this->theme_customize = new VF_Groups_Customize();
+
+    // Add child theme hooks
+    add_filter(
+      'option_blogdescription',
+      array($this, 'option_blogdescription'),
+      10, 1
+    );
+    add_action(
+      'vf/__experimental__/admin/customize',
+      array($this, 'admin_customize')
+    );
+    add_filter(
+      'pre_get_posts',
+      array($this, 'pre_get_posts')
+    );
+    add_filter(
+      'body_class',
+      array($this, 'body_class')
+    );
+    add_action(
+      'wp_head',
+      array($this, 'wp_head')
+    );
+    add_action(
+      'wp_enqueue_scripts',
+      array($this, 'wp_enqueue_scripts')
+    );
+    add_filter(
+      'nav_menu_css_class',
+      array($this, 'nav_menu_css_class'),
+      10, 4
+    );
+    add_filter(
+      'nav_menu_link_attributes',
+      array($this, 'nav_menu_link_attributes'),
+      10, 4
+    );
+  }
+
+  /**
+   * Filter the blog description to load via Content Hub
+   */
+  public function option_blogdescription($value) {
+    // remove filter to avoid update recursion
+    remove_filter(
+      'option_blogdescription',
+      array($this, 'option_blogdescription'),
+      10, 1
+    );
+
+    if ( ! class_exists('VF_Cache')) {
+      return $value;
+    }
+
+    // generate API request
+    $term_id = get_field('embl_taxonomy_term_what', 'option');
+    $uuid = embl_taxonomy_get_uuid($term_id);
+
+    if ( ! $uuid) {
+      return $value;
+    }
+
+    $url = VF_Cache::get_api_url();
+    $url .= '/pattern.html';
+    $url = add_query_arg(array(
+      'filter-uuid'         => $uuid,
+      'filter-content-type' => 'profiles',
+      'pattern'             => 'node-strapline',
+      'source'              => 'contenthub',
+    ), $url);
+
+    // cache for one day
+    $max_age = 60 * 60 * 24 * 1;
+
+    // fetch content via the Content Hub cache
+    $description = VF_Cache::fetch($url, $max_age);
+
+    // strip HTML comments
+    $description = preg_replace('#<!--(.*?)-->#s', '', $description);
+    // strip edit link
+    $description = preg_replace(
+      '#<a[^>]*class="[^"]*embl-conditional-edit[^"]*"[^>]*>.*</a>#s',
+      '', $description
+    );
+    // strip tags except for allowed
+    $description = wp_kses(
+      $description,
+      array(
+        'span' => array()
+      )
+    );
+    $description = trim($description);
+
+    // save updated description
+    if ( ! empty($description) && $value !== $description) {
+      $value = $description;
+      update_option('blogdescription', $value);
+    }
+
+    return $value;
+  }
+
+  /**
+   * Admin Customize
+   * Action: `vf/admin/customize`
+   */
+  public function admin_customize($wp_customize) {
+    // Add setting
+    $wp_customize->add_setting('vf_theme_color', array(
+      'capability' => 'edit_theme_options',
+      'sanitize_callback' => array($this, 'admin_customize_sanitize'),
+      'default' => '009f4d',
+    ));
+    // Add control
+    $wp_customize->add_control('vf_theme_color', array(
+      'type' => 'select',
+      'section' => 'vf_theme',
+      'label' => __('Custom Select Option'),
+      'description' => __('This is a custom select option.'),
+      'choices' => array(
+        '009f4d' => __('EMBL Green'),
+        '007c80' => __('EMBL-EBI Petrol'),
+      ),
+    ));
+  }
+
+  /**
+   * Admin Customize: `sanitize_callback` for select values
+   */
+  public function admin_customize_sanitize($input, $setting) {
+    $input = sanitize_key($input);
+    $choices = $setting->manager->get_control($setting->id)->choices;
+    return (
+      array_key_exists($input, $choices)
+      ? $input
+      : $setting->default
+    );
+  }
+
+  /**
+   * Filter page query
+   */
+  public function pre_get_posts($query) {
+    if (is_admin()) {
+      return $query;
+    }
+    if ( ! $query->is_main_query()) {
+      return $query;
+    }
+    // Exclude non posts from search
+    if ($query->is_search()) {
+      $query->set('post_type', 'post');
+    }
+    return $query;
+  }
+
+  /**
+   * Append <body> class for Visual Framework
+   */
+  public function body_class($classes) {
+    $classes[] = 'vf-body';
+    $classes[] = 'vf-u-background-color-ui--grey';
+    return $classes;
+  }
+
+  /**
+   * Output inline <head> stuff
+   */
+  public function wp_head() {
+    $theme_color = get_theme_mod('vf_theme_color', '009f4d');
+  ?>
+  <style>
+  .vf-wp-theme .vf-box--secondary,
+  .vf-wp-theme .vf-masthead {
+    --vf-masthead__color--background: #<?php echo $theme_color; ?>;
+  }
+  </style>
+  <?php
+  }
+
+  /**
+   * Load CSS and JavaScript assets
+   */
+  public function wp_enqueue_scripts() {
+    $dir = untrailingslashit(get_template_directory_uri());
+    // Register assets for "Jobs" template
+    if (
+      is_page_template('template-jobs.php') &&
+      function_exists('embl_taxonomy')
+    ) {
+      $assets = $dir . '/assets/assets/vfwp-autocomplete';
+      wp_register_script(
+        'vfwp-autocomplete',
+        $assets . '/vfwp-autocomplete.js',
+        array(),
+        '2.0.1',
+        true
+      );
+      wp_register_style(
+        'vfwp-autocomplete',
+        $assets . '/vfwp-autocomplete.css',
+        array(),
+        '2.0.1',
+        'all'
+      );
+      wp_enqueue_script('vfwp-autocomplete');
+      wp_enqueue_style('vfwp-autocomplete');
+    }
+  }
+
+  /**
+   * Add VF class to primary menu items
+   */
+  public function nav_menu_css_class($classes, $item, $args, $depth) {
+    if (in_array($args->theme_location, array('primary', 'secondary'))) {
+      return ['vf-navigation__item'];
+    }
+    return $classes;
+  }
+
+  /**
+   * Add VF class to primary menu items
+   */
+  public function nav_menu_link_attributes($atts, $item, $args, $depth) {
+    if (in_array($args->theme_location, array('primary', 'secondary'))) {
+      $atts['class'] = 'vf-navigation__link';
+    }
+    return $atts;
+  }
+
+} // VF_Groups_Theme
+
+endif;
+
+?>
